Finalizando componentes de formulário e lista de avaliações

+++ AvaliaUNB/src/jvmMain/kotlin/data/models/TeacherModel.kt
@@ -8,6 +8,6 @@ import javax.imageio.ImageIO
 
 data class TeacherModel(
     val name: String,
-    val score: Int?,
+    val score: Float?,
     val profilePicture: ImageBitmap = ImageIO.read(File(ResourcesUtils.ImagePaths.PERSON)).toComposeImageBitmap()
 )
